requirements.txt
pydantic
numpy-financial
jupyter
pandas
matplotlib
google-genai